+++ test/transport/webrtc_transport_test.dart
@@ -3,7 +3,7 @@ import 'dart:typed_data';
 import 'package:test/test.dart';
 import 'package:mockito/mockito.dart';
 import 'package:mockito/annotations.dart';
-import 'package:ipfs_libp2p/dart_libp2p.dart' as libp2p;
+import 'package:dart_libp2p/dart_libp2p.dart' as libp2p;
 import 'package:dart_ipfs/src/transport/webrtc/webrtc_transport.dart';
 import 'package:dart_ipfs/src/transport/webrtc/peer_connection.dart';
 import 'package:dart_ipfs/src/transport/webrtc/data_channel_stream.dart';