+++ test/transport/webtransport_parser_test.dart
@@ -1,5 +1,5 @@
 import 'package:test/test.dart';
-import 'package:ipfs_libp2p/dart_libp2p.dart' as libp2p;
+import 'package:dart_libp2p/dart_libp2p.dart' as libp2p;
 import 'package:dart_ipfs/src/transport/webtransport/multiaddr_parser.dart';
 
 void main() {